+++ touchpad/TouchPad/Objects/TouchPad.build_log.htm
@@ -21,19 +21,34 @@ Target DLL:      STLink\ST-LINKIII-KEIL_SWO.dll V2.0.18.0
 Dialog DLL:      TCM.DLL V1.14.9.0
  
 <h2>Project:</h2>
-I:\TouchPad\TouchPad.uvprojx
-Project File Date:  04/29/2017
+C:\Users\USER\Documents\GitHub\105-2-Embedded_System_Design\touchpad\TouchPad\TouchPad.uvprojx
+Project File Date:  05/28/2017
 
 <h2>Output:</h2>
 *** Using Compiler 'V5.06 update 2 (build 183)', folder: 'C:\Keil_v5\ARM\ARMCC\Bin'
 Build target 'Target 1'
+compiling gpio.c...
+compiling interrupt.c...
 compiling main.c...
-main.c(28): warning:  #111-D: statement is unreachable
+main.c(23): warning:  #111-D: statement is unreachable
   	return 0;
-main.c: 1 warning, 0 errors
+main.c(73): warning:  #1035-D: single-precision operand implicitly converted to double-precision
+  			if(duty < 0.90){
+main.c(74): warning:  #1035-D: single-precision operand implicitly converted to double-precision
+  				duty += 0.1;
+main.c(78): warning:  #1035-D: single-precision operand implicitly converted to double-precision
+  			if(duty > 0.2){
+main.c(79): warning:  #1035-D: single-precision operand implicitly converted to double-precision
+  				duty -= 0.1;
+main.c: 5 warnings, 0 errors
+compiling rcc.c...
+compiling exti.c...
+compiling timer.c...
+assembling startup_stm32f401xe.s...
+compiling system_stm32f4xx.c...
 linking...
-Program Size: Code=864 RO-data=420 RW-data=0 ZI-data=1648  
-".\Objects\TouchPad.axf" - 0 Error(s), 1 Warning(s).
+Program Size: Code=3340 RO-data=436 RW-data=24 ZI-data=1648  
+".\Objects\TouchPad.axf" - 0 Error(s), 5 Warning(s).
 
 <h2>Software Packages used:</h2>
 
@@ -50,7 +65,7 @@ Package Vendor: Keil
    * Component: Startup Version: 2.5.0
 
 <h2>Collection of Component include folders:</h2>
-  I:\TouchPad\RTE
+  C:\Users\USER\Documents\GitHub\105-2-Embedded_System_Design\touchpad\TouchPad\RTE
   C:\Keil_v5\ARM\PACK\ARM\CMSIS\4.5.0\CMSIS\Include
   C:\Keil_v5\ARM\PACK\Keil\STM32F4xx_DFP\2.9.0\Drivers\CMSIS\Device\ST\STM32F4xx\Include
 
@@ -60,11 +75,11 @@ Package Vendor: Keil
 
    * Component: Keil::Device:Startup:2.5.0
       Source file:   MDK\Device\Source\ARM\STM32F4xx_OTP.s
-      Source file:   Drivers\CMSIS\Device\ST\STM32F4xx\Source\Templates\arm\startup_stm32f401xe.s
       Source file:   MDK\Device\Source\ARM\STM32F401xx_OPT.s
+      Source file:   Drivers\CMSIS\Device\ST\STM32F4xx\Source\Templates\arm\startup_stm32f401xe.s
       Include file:  Drivers\CMSIS\Device\ST\STM32F4xx\Include\stm32f4xx.h
       Source file:   Drivers\CMSIS\Device\ST\STM32F4xx\Source\Templates\system_stm32f4xx.c
-Build Time Elapsed:  00:00:00
+Build Time Elapsed:  00:00:02
 </pre>
 </body>
 </html>
