+++ touchpad/TouchPad/Objects/TouchPad.lnp
@@ -1,8 +1,10 @@
 --cpu Cortex-M4.fp
-".\objects\gpio.o"
 ".\objects\main.o"
+".\objects\interrupt.o"
+".\objects\gpio.o"
 ".\objects\rcc.o"
 ".\objects\timer.o"
+".\objects\exti.o"
 ".\objects\startup_stm32f401xe.o"
 ".\objects\system_stm32f4xx.o"
 --ro-base 0x08000000 --entry 0x08000000 --rw-base 0x20000000 --entry Reset_Handler --first __Vectors --strict --summary_stderr --info summarysizes --map --xref --callgraph --symbols
